requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0

# Voice processing
openai>=1.0.0
groq>=0.3.0
sounddevice>=0.4.6
soundfile>=0.12.1

# Video processing
python-obs-websocket>=1.0.0

# Transcription
whisper>=1.0.0
speechrecognition>=3.10.0

# Machine Learning
transformers>=4.36.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# Utilities
python-json-logger>=2.0.7
requests>=2.31.0
aiohttp>=3.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
